anomaly_detector/requirements.txt
celery==5.5.2
redis==6.1.0
python-dotenv
